Nombre del Proyecto: Mi Monte Fase 2 - Conciliación    * Se corrigen defectos encontrados durante inspeccion de CODE [NMPCA-D263, NMPCA-D264, NMPCA-D265, NMPCA-D266]

+++ src/main/java/mx/com/nmp/pagos/mimonte/builder/conciliacion/ComisionesBuilder.java
@@ -98,7 +98,7 @@ public abstract class ComisionesBuilder {
 	 */
 	public static ComisionTransaccion buildComisionTransaccionFromComisionesTransDTOAndComisionesTransaccionesRequestDTO(
 			ComisionesTransDTO comisionesTransDTO, ComisionesTransaccionesRequestDTO comisionesTransaccionesRequestDTO,
-			Integer idConciliacion, String requestUser) {
+			Long idConciliacion, String requestUser) {
 		ComisionTransaccion comisionTransaccion = null;
 		if (null != comisionesTransDTO) {
 			comisionTransaccion = new ComisionTransaccion();
